Add face matching function and video capture loop
This version of the face recognition from video stream program uses the deepface and cv2 (by Open CV) libraries to capture video from the secondary webcam of a laptop and match faces detected in webcam video frames with a reference image.
It displays the webcam video feed back to the user and draws a rectangle around the faces detected in the video frames. The rectangle is green if the detected face matches the face in the reference image, otherwise the rectangle is red.
The webcam feed and face recognition rectangle drawing should stop when the q (for quit) button is pressed.

Issues:
 1 - There is significant lag between displayed frames and the frames recorded by the webcam.
 2 - The rectangle drawn on the faces detected in the webcam video frames are always red, even if the faces detected in the frame are the same face as the reference image.
 3 - Pressing 'q' on the keyboard does not stop the webcam video capture and does not close the window.